Modularize compat src and tests (#2124)
* === BEGIN compat-modules-2 === Move createPortal to its own module (-5 B)

* Move Children into its own module (+0 B)

* Remove dependency on preact src util (-20 B)

* Move memo and PureComponent into their own modules (+0 B)

* Move forwardRef to own module (+18 B)

* Separate event normalization into a separate module (+0 B)

* Separate vnode & attrs compat into render module (-5 B)

* Move isValidElement into its own test file

* Remove duplicate "is exported" tests

* Break out compat index test into a file per function

* Consolidate compat event tests into one file

* Remove jsx.test.js

* Rename createPortal.js to portals.js
