--- compat/test/browser/compat.test.js
@@ -1,61 +1,0 @@
-import { render } from 'preact';
-import { setupScratch, teardown } from '../../../test/_util/helpers';
-
-import React, { createElement } from 'preact/compat';
-
-describe('imported compat in preact', () => {
-	let scratch;
-
-	beforeEach(() => {
-		scratch = setupScratch();
-	});
-
-	afterEach(() => {
-		teardown(scratch);
-	});
-
-	it('should patch events', () => {
-		let spy = sinon.spy();
-		render(<div onClick={spy} />, scratch);
-		scratch.firstChild.click();
-
-		expect(spy).to.be.calledOnce;
-		const event = spy.args[0][0];
-		expect(event).to.haveOwnProperty('persist');
-		expect(event).to.haveOwnProperty('nativeEvent');
-		expect(typeof event.persist).to.equal('function');
-
-		expect(() => event.persist()).to.not.throw();
-	});
-
-	it('should normalize ondoubleclick event', () => {
-		let vnode = <div onDoubleClick={() => null} />;
-		expect(vnode.props).to.haveOwnProperty('ondblclick');
-	});
-
-	it('should normalize onChange for textarea', () => {
-		let vnode = <textarea onChange={() => null} />;
-		expect(vnode.props).to.haveOwnProperty('oninput');
-		expect(vnode.props).to.not.haveOwnProperty('onchange');
-
-		vnode = <textarea oninput={() => null} onChange={() => null} />;
-		expect(vnode.props).to.haveOwnProperty('oninput');
-		expect(vnode.props).to.not.haveOwnProperty('onchange');
-	});
-
-	it('should normalize class+className even on components', () => {
-		function Foo(props) {
-			return (
-				<div class={props.class} className={props.className}>
-					foo
-				</div>
-			);
-		}
-		render(<Foo class="foo" />, scratch);
-		expect(scratch.firstChild.className).to.equal('foo');
-		render(null, scratch);
-
-		render(<Foo className="foo" />, scratch);
-		expect(scratch.firstChild.className).to.equal('foo');
-	});
-});
